Fix dive list rating display and implement consistent tag colors
This comprehensive commit addresses multiple UI/UX improvements for the dive
list and dive sites views, including rating display fixes, tag color
consistency, and UI optimization.

- Fix empty rating display showing '<star> /10' instead of actual ratings
- Update dive list to use 'user_rating' field instead of 'rating' field
- Ensure ratings display correctly in both list and grid view modes

- Create shared tagHelpers.js utility with getTagColor function
- Implement consistent tag color mapping across entire application
- Update dive sites, dive site detail, and dive list views to use shared colors
- Improve visual distinction by changing 'Deep' tag from blue to indigo
- Ensure tags like 'Boat Dive', 'Reef', 'Wreck', 'Deep' have consistent colors

- Add dive tags display to dive list view (previously missing)
- Optimize screen real estate by displaying tags in same row as ratings
- Remove duplicate 'Create Dive Site' button to reduce UI clutter
- Maintain single 'Add Dive Site' button in hero section for better UX

- frontend/src/utils/tagHelpers.js: Shared utility for consistent tag coloring

- frontend/src/pages/Dives.js: Fix rating display, add tags, optimize layout
- frontend/src/pages/DiveSites.js: Use shared tag colors, remove duplicate button
- frontend/src/pages/DiveSiteDetail.js: Use shared tag colors for consistency

- Consistent tag colors across all views (dive sites, dive list, dive detail)
- Better visual distinction between different tag types
- Cleaner interface with reduced redundant UI elements
- Improved information density in dive list view

- All changes tested with Playwright browser automation
- Verified rating display works correctly in dive list
- Confirmed tag colors are consistent across all views
- Validated remaining 'Add Dive Site' button functionality
